layers: Create CoreCheck Queue validation file
