+++ layers/core_checks/core_validation.cpp
@@ -903,555 +903,6 @@ bool CoreChecks::ValidateIndirectCountCmd(const CMD_BUFFER_STATE &cb_state, cons
     return skip;
 }
 
-bool CoreChecks::ValidateCommandBufferSimultaneousUse(const Location &loc, const CMD_BUFFER_STATE &cb_state,
-                                                      int current_submit_count) const {
-    using sync_vuid_maps::GetQueueSubmitVUID;
-    using sync_vuid_maps::SubmitError;
-
-    bool skip = false;
-    if ((cb_state.InUse() || current_submit_count > 1) &&
-        !(cb_state.beginInfo.flags & VK_COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT)) {
-        const auto &vuid = sync_vuid_maps::GetQueueSubmitVUID(loc, SubmitError::kCmdNotSimultaneous);
-
-        skip |= LogError(device, vuid, "%s %s is already in use and is not marked for simultaneous use.", loc.Message().c_str(),
-                         report_data->FormatHandle(cb_state.commandBuffer()).c_str());
-    }
-    return skip;
-}
-
-bool CoreChecks::ValidateCommandBufferState(const CMD_BUFFER_STATE &cb_state, const char *call_source, int current_submit_count,
-                                            const char *vu_id) const {
-    bool skip = false;
-    if (disabled[command_buffer_state]) return skip;
-    // Validate ONE_TIME_SUBMIT_BIT CB is not being submitted more than once
-    if ((cb_state.beginInfo.flags & VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT) &&
-        (cb_state.submitCount + current_submit_count > 1)) {
-        skip |= LogError(cb_state.commandBuffer(), kVUID_Core_DrawState_CommandBufferSingleSubmitViolation,
-                         "%s was begun w/ VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT set, but has been submitted 0x%" PRIxLEAST64
-                         "times.",
-                         report_data->FormatHandle(cb_state.commandBuffer()).c_str(), cb_state.submitCount + current_submit_count);
-    }
-
-    // Validate that cmd buffers have been updated
-    switch (cb_state.state) {
-        case CB_INVALID_INCOMPLETE:
-        case CB_INVALID_COMPLETE:
-            skip |= ReportInvalidCommandBuffer(cb_state, call_source);
-            break;
-
-        case CB_NEW:
-            skip |= LogError(cb_state.commandBuffer(), vu_id, "%s used in the call to %s is unrecorded and contains no commands.",
-                             report_data->FormatHandle(cb_state.commandBuffer()).c_str(), call_source);
-            break;
-
-        case CB_RECORDING:
-            skip |= LogError(cb_state.commandBuffer(), kVUID_Core_DrawState_NoEndCommandBuffer,
-                             "You must call vkEndCommandBuffer() on %s before this call to %s!",
-                             report_data->FormatHandle(cb_state.commandBuffer()).c_str(), call_source);
-            break;
-
-        default: /* recorded */
-            break;
-    }
-    return skip;
-}
-
-// Check that the queue family index of 'queue' matches one of the entries in pQueueFamilyIndices
-bool CoreChecks::ValidImageBufferQueue(const CMD_BUFFER_STATE &cb_state, const VulkanTypedHandle &object, uint32_t queueFamilyIndex,
-                                       uint32_t count, const uint32_t *indices) const {
-    bool found = false;
-    bool skip = false;
-    for (uint32_t i = 0; i < count; i++) {
-        if (indices[i] == queueFamilyIndex) {
-            found = true;
-            break;
-        }
-    }
-
-    if (!found) {
-        const LogObjectList objlist(cb_state.commandBuffer(), object);
-        skip = LogError(objlist, "VUID-vkQueueSubmit-pSubmits-04626",
-                        "vkQueueSubmit: %s contains %s which was not created allowing concurrent access to "
-                        "this queue family %d.",
-                        report_data->FormatHandle(cb_state.commandBuffer()).c_str(), report_data->FormatHandle(object).c_str(),
-                        queueFamilyIndex);
-    }
-    return skip;
-}
-
-// Validate that queueFamilyIndices of primary command buffers match this queue
-// Secondary command buffers were previously validated in vkCmdExecuteCommands().
-bool CoreChecks::ValidateQueueFamilyIndices(const Location &loc, const CMD_BUFFER_STATE &cb_state, VkQueue queue) const {
-    using sync_vuid_maps::GetQueueSubmitVUID;
-    using sync_vuid_maps::SubmitError;
-    bool skip = false;
-    auto pool = cb_state.command_pool;
-    auto queue_state = Get<QUEUE_STATE>(queue);
-
-    if (pool && queue_state) {
-        if (pool->queueFamilyIndex != queue_state->queueFamilyIndex) {
-            const LogObjectList objlist(cb_state.commandBuffer(), queue);
-            const auto &vuid = GetQueueSubmitVUID(loc, SubmitError::kCmdWrongQueueFamily);
-            skip |= LogError(objlist, vuid,
-                             "%s Primary %s created in queue family %d is being submitted on %s "
-                             "from queue family %d.",
-                             loc.Message().c_str(), report_data->FormatHandle(cb_state.commandBuffer()).c_str(),
-                             pool->queueFamilyIndex, report_data->FormatHandle(queue).c_str(), queue_state->queueFamilyIndex);
-        }
-
-        // Ensure that any bound images or buffers created with SHARING_MODE_CONCURRENT have access to the current queue family
-        for (const auto &base_node : cb_state.object_bindings) {
-            switch (base_node->Type()) {
-                case kVulkanObjectTypeImage: {
-                    auto image_state = static_cast<const IMAGE_STATE *>(base_node.get());
-                    if (image_state && image_state->createInfo.sharingMode == VK_SHARING_MODE_CONCURRENT) {
-                        skip |= ValidImageBufferQueue(cb_state, image_state->Handle(), queue_state->queueFamilyIndex,
-                                                      image_state->createInfo.queueFamilyIndexCount,
-                                                      image_state->createInfo.pQueueFamilyIndices);
-                    }
-                    break;
-                }
-                case kVulkanObjectTypeBuffer: {
-                    auto buffer_state = static_cast<const BUFFER_STATE *>(base_node.get());
-                    if (buffer_state && buffer_state->createInfo.sharingMode == VK_SHARING_MODE_CONCURRENT) {
-                        skip |= ValidImageBufferQueue(cb_state, buffer_state->Handle(), queue_state->queueFamilyIndex,
-                                                      buffer_state->createInfo.queueFamilyIndexCount,
-                                                      buffer_state->createInfo.pQueueFamilyIndices);
-                    }
-                    break;
-                }
-                default:
-                    break;
-            }
-        }
-    }
-
-    return skip;
-}
-
-bool CoreChecks::ValidatePrimaryCommandBufferState(
-    const Location &loc, const CMD_BUFFER_STATE &cb_state, int current_submit_count,
-    QFOTransferCBScoreboards<QFOImageTransferBarrier> *qfo_image_scoreboards,
-    QFOTransferCBScoreboards<QFOBufferTransferBarrier> *qfo_buffer_scoreboards) const {
-    using sync_vuid_maps::GetQueueSubmitVUID;
-    using sync_vuid_maps::SubmitError;
-
-    // Track in-use for resources off of primary and any secondary CBs
-    bool skip = false;
-
-    if (cb_state.createInfo.level == VK_COMMAND_BUFFER_LEVEL_SECONDARY) {
-        const auto &vuid = GetQueueSubmitVUID(loc, SubmitError::kSecondaryCmdInSubmit);
-        skip |=
-            LogError(cb_state.commandBuffer(), vuid, "%s Command buffer %s must be allocated with VK_COMMAND_BUFFER_LEVEL_PRIMARY.",
-                     loc.Message().c_str(), report_data->FormatHandle(cb_state.commandBuffer()).c_str());
-    } else {
-        for (const auto *sub_cb : cb_state.linkedCommandBuffers) {
-            skip |= ValidateQueuedQFOTransfers(*sub_cb, qfo_image_scoreboards, qfo_buffer_scoreboards);
-            // TODO: replace with InvalidateCommandBuffers() at recording.
-            if ((sub_cb->primaryCommandBuffer != cb_state.commandBuffer()) &&
-                !(sub_cb->beginInfo.flags & VK_COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT)) {
-                const auto &vuid = GetQueueSubmitVUID(loc, SubmitError::kSecondaryCmdNotSimultaneous);
-                const LogObjectList objlist(device, cb_state.commandBuffer(), sub_cb->commandBuffer(),
-                                            sub_cb->primaryCommandBuffer);
-                skip |= LogError(objlist, vuid,
-                                 "%s %s was submitted with secondary %s but that buffer has subsequently been bound to "
-                                 "primary %s and it does not have VK_COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT set.",
-                                 loc.Message().c_str(), report_data->FormatHandle(cb_state.commandBuffer()).c_str(),
-                                 report_data->FormatHandle(sub_cb->commandBuffer()).c_str(),
-                                 report_data->FormatHandle(sub_cb->primaryCommandBuffer).c_str());
-            }
-
-            if (sub_cb->state != CB_RECORDED) {
-                const char *const finished_cb_vuid = (loc.function == Func::vkQueueSubmit)
-                                                         ? "VUID-vkQueueSubmit-pCommandBuffers-00072"
-                                                         : "VUID-vkQueueSubmit2-commandBuffer-03876";
-                const LogObjectList objlist(device, cb_state.commandBuffer(), sub_cb->commandBuffer(),
-                                            sub_cb->primaryCommandBuffer);
-                skip |= LogError(objlist, finished_cb_vuid,
-                                 "%s: Secondary command buffer %s is not in a valid (pending or executable) state.",
-                                 loc.StringFunc().c_str(), report_data->FormatHandle(sub_cb->commandBuffer()).c_str());
-            }
-        }
-    }
-
-    // If USAGE_SIMULTANEOUS_USE_BIT not set then CB cannot already be executing on device
-    skip |= ValidateCommandBufferSimultaneousUse(loc, cb_state, current_submit_count);
-
-    skip |= ValidateQueuedQFOTransfers(cb_state, qfo_image_scoreboards, qfo_buffer_scoreboards);
-
-    const char *const vuid = (loc.function == Func::vkQueueSubmit) ? "VUID-vkQueueSubmit-pCommandBuffers-00070"
-                                                                   : "VUID-vkQueueSubmit2-commandBuffer-03874";
-    skip |= ValidateCommandBufferState(cb_state, loc.StringFunc().c_str(), current_submit_count, vuid);
-    return skip;
-}
-
-void CoreChecks::PostCallRecordQueueSubmit(VkQueue queue, uint32_t submitCount, const VkSubmitInfo *pSubmits, VkFence fence,
-                                           VkResult result) {
-    StateTracker::PostCallRecordQueueSubmit(queue, submitCount, pSubmits, fence, result);
-
-    if (result != VK_SUCCESS) return;
-    // The triply nested for duplicates that in the StateTracker, but avoids the need for two additional callbacks.
-    for (uint32_t submit_idx = 0; submit_idx < submitCount; submit_idx++) {
-        const VkSubmitInfo *submit = &pSubmits[submit_idx];
-        for (uint32_t i = 0; i < submit->commandBufferCount; i++) {
-            auto cb_state = GetWrite<CMD_BUFFER_STATE>(submit->pCommandBuffers[i]);
-            if (cb_state) {
-                for (auto *secondary_cmd_buffer : cb_state->linkedCommandBuffers) {
-                    UpdateCmdBufImageLayouts(secondary_cmd_buffer);
-                    RecordQueuedQFOTransfers(secondary_cmd_buffer);
-                }
-                UpdateCmdBufImageLayouts(cb_state.get());
-                RecordQueuedQFOTransfers(cb_state.get());
-            }
-        }
-    }
-}
-
-void CoreChecks::RecordQueueSubmit2(VkQueue queue, uint32_t submitCount, const VkSubmitInfo2KHR *pSubmits, VkFence fence,
-                                    VkResult result) {
-    if (result != VK_SUCCESS) return;
-    // The triply nested for duplicates that in the StateTracker, but avoids the need for two additional callbacks.
-    for (uint32_t submit_idx = 0; submit_idx < submitCount; submit_idx++) {
-        const VkSubmitInfo2KHR *submit = &pSubmits[submit_idx];
-        for (uint32_t i = 0; i < submit->commandBufferInfoCount; i++) {
-            auto cb_state = GetWrite<CMD_BUFFER_STATE>(submit->pCommandBufferInfos[i].commandBuffer);
-            if (cb_state) {
-                for (auto *secondaryCmdBuffer : cb_state->linkedCommandBuffers) {
-                    UpdateCmdBufImageLayouts(secondaryCmdBuffer);
-                    RecordQueuedQFOTransfers(secondaryCmdBuffer);
-                }
-                UpdateCmdBufImageLayouts(cb_state.get());
-                RecordQueuedQFOTransfers(cb_state.get());
-            }
-        }
-    }
-}
-
-void CoreChecks::PostCallRecordQueueSubmit2KHR(VkQueue queue, uint32_t submitCount, const VkSubmitInfo2KHR *pSubmits, VkFence fence,
-                                               VkResult result) {
-    StateTracker::PostCallRecordQueueSubmit2KHR(queue, submitCount, pSubmits, fence, result);
-    RecordQueueSubmit2(queue, submitCount, pSubmits, fence, result);
-}
-
-void CoreChecks::PostCallRecordQueueSubmit2(VkQueue queue, uint32_t submitCount, const VkSubmitInfo2 *pSubmits, VkFence fence,
-                                            VkResult result) {
-    StateTracker::PostCallRecordQueueSubmit2(queue, submitCount, pSubmits, fence, result);
-    RecordQueueSubmit2(queue, submitCount, pSubmits, fence, result);
-}
-
-struct CommandBufferSubmitState {
-    const CoreChecks *core;
-    const QUEUE_STATE *queue_state;
-    QFOTransferCBScoreboards<QFOImageTransferBarrier> qfo_image_scoreboards;
-    QFOTransferCBScoreboards<QFOBufferTransferBarrier> qfo_buffer_scoreboards;
-    vector<VkCommandBuffer> current_cmds;
-    GlobalImageLayoutMap overlay_image_layout_map;
-    QueryMap local_query_to_state_map;
-    EventToStageMap local_event_to_stage_map;
-
-    CommandBufferSubmitState(const CoreChecks *c, const char *func, const QUEUE_STATE *q) : core(c), queue_state(q) {}
-
-    bool Validate(const core_error::Location &loc, const CMD_BUFFER_STATE &cb_state, uint32_t perf_pass) {
-        bool skip = false;
-        skip |= core->ValidateCmdBufImageLayouts(loc, cb_state, overlay_image_layout_map);
-        auto cmd = cb_state.commandBuffer();
-        current_cmds.push_back(cmd);
-        skip |= core->ValidatePrimaryCommandBufferState(loc, cb_state,
-                                                        static_cast<int>(std::count(current_cmds.begin(), current_cmds.end(), cmd)),
-                                                        &qfo_image_scoreboards, &qfo_buffer_scoreboards);
-        skip |= core->ValidateQueueFamilyIndices(loc, cb_state, queue_state->Queue());
-
-        for (const auto &descriptor_set : cb_state.validate_descriptorsets_in_queuesubmit) {
-            auto set_node = core->Get<cvdescriptorset::DescriptorSet>(descriptor_set.first);
-            if (!set_node) {
-                continue;
-            }
-            for (const auto &cmd_info : descriptor_set.second) {
-                // dynamic data isn't allowed in UPDATE_AFTER_BIND, so dynamicOffsets is always empty.
-                std::vector<uint32_t> dynamic_offsets;
-                std::optional<vvl::unordered_map<VkImageView, VkImageLayout>> checked_layouts;
-
-                std::string function = loc.StringFunc();
-                function += ", ";
-                function += CommandTypeString(cmd_info.cmd_type);
-                CoreChecks::DescriptorContext context{function.c_str(),
-                                                      core->GetDrawDispatchVuid(cmd_info.cmd_type),
-                                                      cb_state,
-                                                      *set_node,
-                                                      cmd_info.framebuffer,
-                                                      false,  // This is submit time not record time...
-                                                      dynamic_offsets,
-                                                      checked_layouts};
-
-                for (const auto &binding_info : cmd_info.binding_infos) {
-                    std::string error;
-                    if (set_node->GetTotalDescriptorCount() > cvdescriptorset::PrefilterBindRequestMap::kManyDescriptors_) {
-                        context.checked_layouts.emplace();
-                    }
-                    const auto *binding = set_node->GetBinding(binding_info.first);
-                    skip |= core->ValidateDescriptorSetBindingData(context, binding_info, *binding);
-                }
-            }
-        }
-
-        // Potential early exit here as bad object state may crash in delayed function calls
-        if (skip) {
-            return true;
-        }
-
-        // Call submit-time functions to validate or update local mirrors of state (to preserve const-ness at validate time)
-        for (auto &function : cb_state.queue_submit_functions) {
-            skip |= function(*core, *queue_state, cb_state);
-        }
-        for (auto &function : cb_state.eventUpdates) {
-            skip |= function(const_cast<CMD_BUFFER_STATE &>(cb_state), /*do_validate*/ true, &local_event_to_stage_map);
-        }
-        VkQueryPool first_perf_query_pool = VK_NULL_HANDLE;
-        for (auto &function : cb_state.queryUpdates) {
-            skip |= function(const_cast<CMD_BUFFER_STATE &>(cb_state), /*do_validate*/ true, first_perf_query_pool, perf_pass,
-                             &local_query_to_state_map);
-        }
-        for (const auto &it : cb_state.video_session_updates) {
-            auto video_session_state = core->Get<VIDEO_SESSION_STATE>(it.first);
-            VideoSessionDeviceState local_state = video_session_state->DeviceStateCopy();
-            for (const auto &function : it.second) {
-                skip |= function(core, video_session_state.get(), local_state, /*do_validate*/ true);
-            }
-        }
-        return skip;
-    }
-};
-
-bool CoreChecks::PreCallValidateQueueSubmit(VkQueue queue, uint32_t submitCount, const VkSubmitInfo *pSubmits,
-                                            VkFence fence) const {
-    auto fence_state = Get<FENCE_STATE>(fence);
-    bool skip = ValidateFenceForSubmit(fence_state.get(), "VUID-vkQueueSubmit-fence-00064", "VUID-vkQueueSubmit-fence-00063",
-                                       "vkQueueSubmit()");
-    if (skip) {
-        return true;
-    }
-    auto queue_state = Get<QUEUE_STATE>(queue);
-    CommandBufferSubmitState cb_submit_state(this, "vkQueueSubmit()", queue_state.get());
-    SemaphoreSubmitState sem_submit_state(this, queue,
-                                          physical_device_state->queue_family_properties[queue_state->queueFamilyIndex].queueFlags);
-
-    // Now verify each individual submit
-    for (uint32_t submit_idx = 0; submit_idx < submitCount; submit_idx++) {
-        const VkSubmitInfo &submit = pSubmits[submit_idx];
-        const auto perf_submit = LvlFindInChain<VkPerformanceQuerySubmitInfoKHR>(submit.pNext);
-        uint32_t perf_pass = perf_submit ? perf_submit->counterPassIndex : 0;
-
-        Location loc(Func::vkQueueSubmit, Struct::VkSubmitInfo, Field::pSubmits, submit_idx);
-        bool suspended_render_pass_instance = false;
-        for (uint32_t i = 0; i < submit.commandBufferCount; i++) {
-            auto cb_state = GetRead<CMD_BUFFER_STATE>(submit.pCommandBuffers[i]);
-            if (cb_state) {
-                skip |= cb_submit_state.Validate(loc.dot(Field::pCommandBuffers, i), *cb_state, perf_pass);
-
-                // Validate flags for dynamic rendering
-                if (suspended_render_pass_instance && cb_state->hasRenderPassInstance && !cb_state->resumesRenderPassInstance) {
-                    skip |= LogError(queue, "VUID-VkSubmitInfo-pCommandBuffers-06016",
-                                     "pSubmits[%" PRIu32 "] has a suspended render pass instance, but pCommandBuffers[%" PRIu32
-                                     "] has its own render pass instance that does not resume it.",
-                                     submit_idx, i);
-                }
-                if (cb_state->resumesRenderPassInstance) {
-                    if (!suspended_render_pass_instance) {
-                        skip |= LogError(queue, "VUID-VkSubmitInfo-pCommandBuffers-06193",
-                                         "pSubmits[%" PRIu32 "]->pCommandBuffers[%" PRIu32
-                                         "] resumes a render pass instance, but there is no suspended render pass instance.",
-                                         submit_idx, i);
-                    }
-                    suspended_render_pass_instance = false;
-                }
-                if (cb_state->suspendsRenderPassInstance) {
-                    suspended_render_pass_instance = true;
-                }
-            }
-        }
-        // Renderpass should not be in suspended state after the final cmdbuf
-        if (suspended_render_pass_instance) {
-            skip |= LogError(queue, "VUID-VkSubmitInfo-pCommandBuffers-06014",
-                             "pSubmits[%" PRIu32 "] has a suspended render pass instance that was not resumed.", submit_idx);
-        }
-        skip |= ValidateSemaphoresForSubmit(sem_submit_state, submit, loc);
-
-        auto chained_device_group_struct = LvlFindInChain<VkDeviceGroupSubmitInfo>(submit.pNext);
-        if (chained_device_group_struct && chained_device_group_struct->commandBufferCount > 0) {
-            for (uint32_t i = 0; i < chained_device_group_struct->commandBufferCount; ++i) {
-                const LogObjectList objlist(queue);
-                skip |= ValidateDeviceMaskToPhysicalDeviceCount(chained_device_group_struct->pCommandBufferDeviceMasks[i], objlist,
-                                                                "VUID-VkDeviceGroupSubmitInfo-pCommandBufferDeviceMasks-00086");
-            }
-            if (chained_device_group_struct->signalSemaphoreCount != submit.signalSemaphoreCount) {
-                skip |= LogError(queue, "VUID-VkDeviceGroupSubmitInfo-signalSemaphoreCount-00084",
-                                 "pSubmits[%" PRIu32 "] signalSemaphoreCount (%" PRIu32
-                                 ") is different than signalSemaphoreCount (%" PRIu32
-                                 ") of the VkDeviceGroupSubmitInfo in its pNext chain",
-                                 submit_idx, submit.signalSemaphoreCount, chained_device_group_struct->signalSemaphoreCount);
-            }
-            if (chained_device_group_struct->waitSemaphoreCount != submit.waitSemaphoreCount) {
-                skip |=
-                    LogError(queue, "VUID-VkDeviceGroupSubmitInfo-waitSemaphoreCount-00082",
-                             "pSubmits[%" PRIu32 "] waitSemaphoreCount (%" PRIu32 ") is different than waitSemaphoreCount (%" PRIu32
-                             ") of the VkDeviceGroupSubmitInfo in its pNext chain",
-                             submit_idx, submit.waitSemaphoreCount, chained_device_group_struct->waitSemaphoreCount);
-            }
-            if (chained_device_group_struct->commandBufferCount != submit.commandBufferCount) {
-                skip |=
-                    LogError(queue, "VUID-VkDeviceGroupSubmitInfo-commandBufferCount-00083",
-                             "pSubmits[%" PRIu32 "] commandBufferCount (%" PRIu32 ") is different than commandBufferCount (%" PRIu32
-                             ") of the VkDeviceGroupSubmitInfo in its pNext chain",
-                             submit_idx, submit.commandBufferCount, chained_device_group_struct->commandBufferCount);
-            }
-        }
-
-        auto protected_submit_info = LvlFindInChain<VkProtectedSubmitInfo>(submit.pNext);
-        if (protected_submit_info) {
-            const bool protected_submit = protected_submit_info->protectedSubmit == VK_TRUE;
-            if ((protected_submit == true) && ((queue_state->flags & VK_DEVICE_QUEUE_CREATE_PROTECTED_BIT) == 0)) {
-                skip |= LogError(queue, "VUID-vkQueueSubmit-queue-06448",
-                                 "vkQueueSubmit(): pSubmits[%u] contains a protected submission to %s which was not created with "
-                                 "VK_DEVICE_QUEUE_CREATE_PROTECTED_BIT",
-                                 submit_idx, report_data->FormatHandle(queue).c_str());
-            }
-
-            // Make sure command buffers are all protected or unprotected
-            for (uint32_t i = 0; i < submit.commandBufferCount; i++) {
-                auto cb_state = GetRead<CMD_BUFFER_STATE>(submit.pCommandBuffers[i]);
-                if (cb_state) {
-                    if ((cb_state->unprotected == true) && (protected_submit == true)) {
-                        const LogObjectList objlist(cb_state->commandBuffer(), queue);
-                        skip |= LogError(objlist, "VUID-VkSubmitInfo-pNext-04148",
-                                         "vkQueueSubmit(): command buffer %s is unprotected while queue %s pSubmits[%u] has "
-                                         "VkProtectedSubmitInfo:protectedSubmit set to VK_TRUE",
-                                         report_data->FormatHandle(cb_state->commandBuffer()).c_str(),
-                                         report_data->FormatHandle(queue).c_str(), submit_idx);
-                    }
-                    if ((cb_state->unprotected == false) && (protected_submit == false)) {
-                        const LogObjectList objlist(cb_state->commandBuffer(), queue);
-                        skip |= LogError(objlist, "VUID-VkSubmitInfo-pNext-04120",
-                                         "vkQueueSubmit(): command buffer %s is protected while queue %s pSubmits[%u] has "
-                                         "VkProtectedSubmitInfo:protectedSubmit set to VK_FALSE",
-                                         report_data->FormatHandle(cb_state->commandBuffer()).c_str(),
-                                         report_data->FormatHandle(queue).c_str(), submit_idx);
-                    }
-                }
-            }
-        }
-    }
-
-    return skip;
-}
-
-bool CoreChecks::ValidateQueueSubmit2(VkQueue queue, uint32_t submitCount, const VkSubmitInfo2KHR *pSubmits, VkFence fence,
-                                      bool is_2khr) const {
-    auto pFence = Get<FENCE_STATE>(fence);
-    const char *func_name = is_2khr ? "vkQueueSubmit2KHR()" : "vkQueueSubmit2()";
-    bool skip =
-        ValidateFenceForSubmit(pFence.get(), "VUID-vkQueueSubmit2-fence-04895", "VUID-vkQueueSubmit2-fence-04894", func_name);
-    if (skip) {
-        return true;
-    }
-
-    if (!enabled_features.core13.synchronization2) {
-        skip |=
-            LogError(queue, "VUID-vkQueueSubmit2-synchronization2-03866", "%s: Synchronization2 feature is not enabled", func_name);
-    }
-
-    auto queue_state = Get<QUEUE_STATE>(queue);
-    CommandBufferSubmitState cb_submit_state(this, func_name, queue_state.get());
-    SemaphoreSubmitState sem_submit_state(this, queue,
-                                          physical_device_state->queue_family_properties[queue_state->queueFamilyIndex].queueFlags);
-
-    // Now verify each individual submit
-    for (uint32_t submit_idx = 0; submit_idx < submitCount; submit_idx++) {
-        const VkSubmitInfo2KHR &submit = pSubmits[submit_idx];
-        const auto perf_submit = LvlFindInChain<VkPerformanceQuerySubmitInfoKHR>(submit.pNext);
-        uint32_t perf_pass = perf_submit ? perf_submit->counterPassIndex : 0;
-        Location loc(Func::vkQueueSubmit2, Struct::VkSubmitInfo2, Field::pSubmits, submit_idx);
-
-        skip |= ValidateSemaphoresForSubmit(sem_submit_state, submit, loc);
-
-        const bool protected_submit = (submit.flags & VK_SUBMIT_PROTECTED_BIT_KHR) != 0;
-        if ((protected_submit == true) && ((queue_state->flags & VK_DEVICE_QUEUE_CREATE_PROTECTED_BIT)) == 0) {
-            skip |= LogError(queue, "VUID-vkQueueSubmit2-queue-06447",
-                             "%s: pSubmits[%u] contains a protected submission to %s which was not created with "
-                             "VK_DEVICE_QUEUE_CREATE_PROTECTED_BIT",
-                             func_name, submit_idx, report_data->FormatHandle(queue).c_str());
-        }
-
-        bool suspended_render_pass_instance = false;
-        for (uint32_t i = 0; i < submit.commandBufferInfoCount; i++) {
-            auto info_loc = loc.dot(Field::pCommandBufferInfos, i);
-            info_loc.structure = Struct::VkCommandBufferSubmitInfo;
-            auto cb_state = GetRead<CMD_BUFFER_STATE>(submit.pCommandBufferInfos[i].commandBuffer);
-            skip |= cb_submit_state.Validate(info_loc.dot(Field::commandBuffer), *cb_state, perf_pass);
-
-            {
-                const LogObjectList objlist(queue);
-                skip |= ValidateDeviceMaskToPhysicalDeviceCount(submit.pCommandBufferInfos[i].deviceMask, queue,
-                                                                "VUID-VkCommandBufferSubmitInfo-deviceMask-03891");
-            }
-
-            if (cb_state != nullptr) {
-                // Make sure command buffers are all protected or unprotected
-                if ((cb_state->unprotected == true) && (protected_submit == true)) {
-                    const LogObjectList objlist(cb_state->commandBuffer(), queue);
-                    skip |= LogError(objlist, "VUID-VkSubmitInfo2-flags-03886",
-                                     "%s: command buffer %s is unprotected while queue %s pSubmits[%u] has "
-                                     "VK_SUBMIT_PROTECTED_BIT_KHR set",
-                                     func_name, report_data->FormatHandle(cb_state->commandBuffer()).c_str(),
-                                     report_data->FormatHandle(queue).c_str(), submit_idx);
-                }
-                if ((cb_state->unprotected == false) && (protected_submit == false)) {
-                    const LogObjectList objlist(cb_state->commandBuffer(), queue);
-                    skip |= LogError(objlist, "VUID-VkSubmitInfo2-flags-03887",
-                                     "%s: command buffer %s is protected while queue %s pSubmitInfos[%u] has "
-                                     "VK_SUBMIT_PROTECTED_BIT_KHR not set",
-                                     func_name, report_data->FormatHandle(cb_state->commandBuffer()).c_str(),
-                                     report_data->FormatHandle(queue).c_str(), submit_idx);
-                }
-
-                if (suspended_render_pass_instance && cb_state->hasRenderPassInstance && !cb_state->resumesRenderPassInstance) {
-                    skip |= LogError(queue, "VUID-VkSubmitInfo2KHR-commandBuffer-06012",
-                                     "pSubmits[%" PRIu32 "] has a suspended render pass instance, but pCommandBuffers[%" PRIu32
-                                     "] has its own render pass instance that does not resume it.",
-                                     submit_idx, i);
-                }
-                if (cb_state->suspendsRenderPassInstance) {
-                    suspended_render_pass_instance = true;
-                }
-                if (cb_state->resumesRenderPassInstance) {
-                    if (!suspended_render_pass_instance) {
-                        skip |= LogError(queue, "VUID-VkSubmitInfo2KHR-commandBuffer-06192",
-                                         "pSubmits[%" PRIu32 "]->pCommandBuffers[%" PRIu32
-                                         "] resumes a render pass instance, but there is no suspended render pass instance.",
-                                         submit_idx, i);
-                    }
-                    suspended_render_pass_instance = false;
-                }
-            }
-        }
-        if (suspended_render_pass_instance) {
-            skip |= LogError(queue, "VUID-VkSubmitInfo2KHR-commandBuffer-06010",
-                             "pSubmits[%" PRIu32 "] has a suspended render pass instance that was not resumed.", submit_idx);
-        }
-    }
-
-    return skip;
-}
-
-bool CoreChecks::PreCallValidateQueueSubmit2KHR(VkQueue queue, uint32_t submitCount, const VkSubmitInfo2KHR *pSubmits,
-                                                VkFence fence) const {
-    return ValidateQueueSubmit2(queue, submitCount, pSubmits, fence, true);
-}
-
-bool CoreChecks::PreCallValidateQueueSubmit2(VkQueue queue, uint32_t submitCount, const VkSubmitInfo2 *pSubmits,
-                                             VkFence fence) const {
-    return ValidateQueueSubmit2(queue, submitCount, pSubmits, fence, false);
-}
-
 // For given obj node, if it is use, flag a validation error and return callback result, else return false
 bool CoreChecks::ValidateObjectNotInUse(const BASE_NODE *obj_node, const char *caller_name, const char *error_code) const {
     if (disabled[object_in_use]) return false;
@@ -2974,91 +2425,6 @@ bool CoreChecks::PreCallValidateCmdExecuteCommands(VkCommandBuffer commandBuffer
     }
 
     skip |= ValidateCmd(*cb_state, CMD_EXECUTECOMMANDS);
-    return skip;
-}
-
-bool CoreChecks::PreCallValidateQueueBindSparse(VkQueue queue, uint32_t bindInfoCount, const VkBindSparseInfo *pBindInfo,
-                                                VkFence fence) const {
-    auto queue_data = Get<QUEUE_STATE>(queue);
-    auto fence_state = Get<FENCE_STATE>(fence);
-    bool skip = ValidateFenceForSubmit(fence_state.get(), "VUID-vkQueueBindSparse-fence-01114",
-                                       "VUID-vkQueueBindSparse-fence-01113", "VkQueueBindSparse()");
-    if (skip) {
-        return true;
-    }
-
-    const auto queue_flags = physical_device_state->queue_family_properties[queue_data->queueFamilyIndex].queueFlags;
-    if (!(queue_flags & VK_QUEUE_SPARSE_BINDING_BIT)) {
-        skip |= LogError(queue, "VUID-vkQueueBindSparse-queuetype",
-                         "vkQueueBindSparse(): a non-memory-management capable queue -- VK_QUEUE_SPARSE_BINDING_BIT not set.");
-    }
-
-    SemaphoreSubmitState sem_submit_state(this, queue,
-                                          physical_device_state->queue_family_properties[queue_data->queueFamilyIndex].queueFlags);
-    for (uint32_t bind_idx = 0; bind_idx < bindInfoCount; ++bind_idx) {
-        Location loc(Func::vkQueueBindSparse, Struct::VkBindSparseInfo);
-        const VkBindSparseInfo &bind_info = pBindInfo[bind_idx];
-
-        skip |= ValidateSemaphoresForSubmit(sem_submit_state, bind_info, loc);
-
-        if (bind_info.pBufferBinds) {
-            for (uint32_t buffer_idx = 0; buffer_idx < bind_info.bufferBindCount; ++buffer_idx) {
-                const VkSparseBufferMemoryBindInfo &buffer_bind = bind_info.pBufferBinds[buffer_idx];
-                if (buffer_bind.pBinds) {
-                    auto buffer_state = Get<BUFFER_STATE>(buffer_bind.buffer);
-                    for (uint32_t buffer_bind_idx = 0; buffer_bind_idx < buffer_bind.bindCount; ++buffer_bind_idx) {
-                        const VkSparseMemoryBind &memory_bind = buffer_bind.pBinds[buffer_bind_idx];
-                        std::stringstream parameter_name;
-                        parameter_name << "pBindInfo[" << bind_idx << "].pBufferBinds[" << buffer_idx << " ].pBinds["
-                                       << buffer_bind_idx << "]";
-                        skip |= ValidateSparseMemoryBind(memory_bind, buffer_state->requirements.size, "vkQueueBindSparse()",
-                                                         parameter_name.str().c_str());
-                    }
-                }
-            }
-        }
-
-        if (bind_info.pImageOpaqueBinds) {
-            for (uint32_t image_opaque_idx = 0; image_opaque_idx < bind_info.bufferBindCount; ++image_opaque_idx) {
-                const VkSparseImageOpaqueMemoryBindInfo &image_opaque_bind = bind_info.pImageOpaqueBinds[image_opaque_idx];
-                if (image_opaque_bind.pBinds) {
-                    auto image_state = Get<IMAGE_STATE>(image_opaque_bind.image);
-                    for (uint32_t image_opaque_bind_idx = 0; image_opaque_bind_idx < image_opaque_bind.bindCount;
-                         ++image_opaque_bind_idx) {
-                        const VkSparseMemoryBind &memory_bind = image_opaque_bind.pBinds[image_opaque_bind_idx];
-                        std::stringstream parameter_name;
-                        parameter_name << "pBindInfo[" << bind_idx << "].pImageOpaqueBinds[" << image_opaque_idx << " ].pBinds["
-                                       << image_opaque_bind_idx << "]";
-                        // Assuming that no multiplanar disjointed images are possible with sparse memory binding. Needs
-                        // confirmation
-                        skip |= ValidateSparseMemoryBind(memory_bind, image_state->requirements[0].size, "vkQueueBindSparse()",
-                                                         parameter_name.str().c_str());
-                    }
-                }
-            }
-        }
-
-        if (bind_info.pImageBinds) {
-            for (uint32_t image_idx = 0; image_idx < bind_info.imageBindCount; ++image_idx) {
-                const VkSparseImageMemoryBindInfo &image_bind = bind_info.pImageBinds[image_idx];
-                auto image_state = Get<IMAGE_STATE>(image_bind.image);
-
-                if (image_state && !(image_state->createInfo.flags & VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT)) {
-                    skip |= LogError(image_bind.image, "VUID-VkSparseImageMemoryBindInfo-image-02901",
-                                     "vkQueueBindSparse(): pBindInfo[%u].pImageBinds[%u]: image must have been created with "
-                                     "VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT set",
-                                     bind_idx, image_idx);
-                }
-
-                if (image_bind.pBinds) {
-                    for (uint32_t image_bind_idx = 0; image_bind_idx < image_bind.bindCount; ++image_bind_idx) {
-                        const VkSparseImageMemoryBind &memory_bind = image_bind.pBinds[image_bind_idx];
-                        skip |= ValidateSparseImageMemoryBind(image_state.get(), memory_bind, image_idx, image_bind_idx);
-                    }
-                }
-            }
-        }
-    }
     return skip;
 }
